+++ cli/commands/asset/list_test.go
@@ -1,6 +1,7 @@
 package asset
 
 import (
+	"encoding/json"
 	"errors"
 	"net/http"
 	"testing"
@@ -159,4 +160,85 @@ func TestListCommandRunEClosureWithHeader(t *testing.T) {
 	assert.Nil(err)
 	assert.Contains(out, "E_TOO_MANY_ENTITIES")
 	assert.Contains(out, "==")
+}
+
+func TestListBuildsWithTabular(t *testing.T) {
+	assert := assert.New(t)
+
+	cli := test.NewMockCLI()
+	config := cli.Config.(*client.MockConfig)
+	config.On("Format").Return("tabular")
+
+	asset := *corev2.FixtureAsset("builds")
+	asset.Builds = []*corev2.AssetBuild{
+		&corev2.AssetBuild{
+			URL:    "http://127.0.0.1/foo",
+			Sha512: "25e01b962045f4f5b624c3e47e782bef65c6c82602524dc569a8431b76cc1f57639d267380a7ec49f70876339ae261704fc51ed2fc520513cf94bc45ed7f6e17",
+		},
+		&corev2.AssetBuild{
+			URL:    "http://127.0.0.1/bar",
+			Sha512: "25e01b962045f4f5b624c3e47e782bef65c6c82602524dc569a8431b76cc1f57639d267380a7ec49f70876339ae261704fc51ed2fc520513cf94bc45ed7f6e17",
+		},
+	}
+
+	client := cli.Client.(*client.MockClient)
+	var header http.Header
+	resources := []corev2.Asset{}
+	client.On("List", mock.Anything, &resources, mock.Anything, &header).Return(nil).Run(
+		func(args mock.Arguments) {
+			resources := args[1].(*[]corev2.Asset)
+			*resources = []corev2.Asset{asset}
+		},
+	)
+
+	cmd := ListCommand(cli)
+	out, err := test.RunCmd(cmd, []string{})
+
+	assert.NotEmpty(out)
+	assert.Nil(err)
+
+	// Make sure the asset's builds have been seperated into several assets
+	assert.Contains(out, "//127.0.0.1/.../foo")
+	assert.Contains(out, "//127.0.0.1/.../bar")
+}
+
+func TestListBuildsWithJSON(t *testing.T) {
+	assert := assert.New(t)
+
+	cli := test.NewMockCLI()
+	config := cli.Config.(*client.MockConfig)
+	config.On("Format").Return("json")
+
+	asset := *corev2.FixtureAsset("builds")
+	asset.Builds = []*corev2.AssetBuild{
+		&corev2.AssetBuild{
+			URL:    "http://127.0.0.1/foo",
+			Sha512: "25e01b962045f4f5b624c3e47e782bef65c6c82602524dc569a8431b76cc1f57639d267380a7ec49f70876339ae261704fc51ed2fc520513cf94bc45ed7f6e17",
+		},
+		&corev2.AssetBuild{
+			URL:    "http://127.0.0.1/bar",
+			Sha512: "25e01b962045f4f5b624c3e47e782bef65c6c82602524dc569a8431b76cc1f57639d267380a7ec49f70876339ae261704fc51ed2fc520513cf94bc45ed7f6e17",
+		},
+	}
+
+	client := cli.Client.(*client.MockClient)
+	var header http.Header
+	resources := []corev2.Asset{}
+	client.On("List", mock.Anything, &resources, mock.Anything, &header).Return(nil).Run(
+		func(args mock.Arguments) {
+			resources := args[1].(*[]corev2.Asset)
+			*resources = []corev2.Asset{asset}
+		},
+	)
+
+	cmd := ListCommand(cli)
+	out, err := test.RunCmd(cmd, []string{})
+
+	assert.NotEmpty(out)
+	assert.Nil(err)
+
+	// Make sure the asset's builds have *not* been seperated into several assets
+	output := []map[string]interface{}{}
+	_ = json.Unmarshal([]byte(out), &output)
+	assert.Len(output, 1)
 }